Zadanie 12. Formatowanie calego kodu. Usunieto Node.java bo jest to obiekt wew listy i nie musi istniec samodzielnie. Niej jest bezposrednio dostepny z API listy
