+++ src/main/java/edu/kis/vh/nursery/HanoiRyhmer.java
@@ -8,6 +8,7 @@ public class HanoiRyhmer extends DefaultCountingOutRyhmer {
 		return totalRejected;
 	}
 
+	@Override
 	public void countIn(int in) {
 		if (!callCheck() && in > peekaboo())
 			totalRejected++;
